Export Overhaul
Basically I missunderstood exports.
Originally I had exported functions defined in the exports using other functions (which is fine and works), but I forgot  to put in the return statements leading me down a rabit hole of debugging undefined return values. It is there that I found a simpler way of exporting (that I'm pretty sure works). Now it should just simply export the already defined functions as is.