Added pixel dimensions fields to Photos Document
This is fulfilling a request from the front end. This adds pixel height and pixel width as fields to the Photos Documents in the Database.

Effected functions:
    Added:
        getPhotoDimensions(userID, photoID) -> {height:val, width:val}
        updatePhotoDimensions(userID, photoID, height, width) -> Boolean

    changed:
        addPhoto(userID, photoReference, photoDateTime=null) to
        addPhoto(userID, photoReference, height, width, photoDateTime=null)

Other functions elsewhere in the database may need to be updated. Backend Routes and Util are yet to be updated to accomidated the change.