Another big batch of changes, and a compromise that can probably be fixed pretty easily:
MainPage.js - Deleted the sidebar to be moved to the correct file
Sidebar.js - implemented feature
api.js - restructured AddImagesToAlbum and CreateNewAlbum functions as they were previously in a broken state. Current implementation will wait for each thing to be added before moving on to the next one as the back-end route for adding many images cannot currently handle sending an array of images. This should be pretty easy to fix, but the current implementation works for now.